Examples/Vertexing: reorganize executables and replace read-based example
