+++ Examples/Run/CMakeLists.txt
@@ -14,4 +14,4 @@ add_subdirectory(Misc)
 add_subdirectory(Propagation)
 add_subdirectory(ReadCsv)
 add_subdirectory(Reconstruction)
-add_subdirectory_if(Vertexing ACTS_BUILD_EXAMPLES_PYTHIA8)
+add_subdirectory(Vertexing)
